+++ kadai-adapter-camunda-spring-boot-test/src/test/java/io/kadai/adapter/integration/TestTaskAcquisition.java
@@ -21,7 +21,7 @@ package io.kadai.adapter.integration;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-import io.kadai.adapter.impl.KadaiTaskStarterOrchestrator;
+import io.kadai.adapter.impl.scheduled.KadaiTaskStarterOrchestrator;
 import io.kadai.adapter.manager.AdapterManager;
 import io.kadai.adapter.systemconnector.api.SystemConnector;
 import io.kadai.adapter.systemconnector.camunda.api.impl.CamundaSystemConnectorImpl;
