+++ kadai-adapter/src/main/java/io/kadai/adapter/impl/scheduled/KadaiTaskCompletionOrchestrator.java
@@ -16,9 +16,10 @@
  *
  */
 
-package io.kadai.adapter.impl;
+package io.kadai.adapter.impl.scheduled;
 
 import io.kadai.adapter.exceptions.TaskTerminationFailedException;
+import io.kadai.adapter.impl.service.KadaiTaskCompletionService;
 import io.kadai.adapter.manager.AdapterManager;
 import io.kadai.adapter.systemconnector.api.ReferencedTask;
 import io.kadai.adapter.systemconnector.api.SystemConnector;
