+++ kadai-adapter/src/main/java/io/kadai/adapter/impl/scheduled/KadaiTaskStarterOrchestrator.java
@@ -16,9 +16,10 @@
  *
  */
 
-package io.kadai.adapter.impl;
+package io.kadai.adapter.impl.scheduled;
 
 import io.kadai.adapter.exceptions.TaskCreationFailedException;
+import io.kadai.adapter.impl.service.KadaiTaskStarterService;
 import io.kadai.adapter.manager.AdapterManager;
 import io.kadai.adapter.systemconnector.api.ReferencedTask;
 import io.kadai.adapter.systemconnector.api.SystemConnector;
