+++ kadai-adapter/src/main/java/io/kadai/adapter/impl/scheduled/SchedulerRun.java
@@ -1,4 +1,4 @@
-package io.kadai.adapter.impl;
+package io.kadai.adapter.impl.scheduled;
 
 import java.time.Instant;
