+++ kadai-adapter/src/main/java/io/kadai/adapter/impl/service/KadaiTaskStarterServiceImpl.java
@@ -16,7 +16,7 @@
  *
  */
 
-package io.kadai.adapter.impl;
+package io.kadai.adapter.impl.service;
 
 import io.kadai.adapter.exceptions.TaskCreationFailedException;
 import io.kadai.adapter.kadaiconnector.api.KadaiConnector;
